move LikeliHood Classes into Fit classes
