+++ src/main/java/fact/features/muon/GaussianFit.java
@@ -3,6 +3,7 @@ package fact.features.muon;
 import fact.Constants;
 import fact.hexmap.FactPixelMapping;
 import fact.hexmap.ui.overlays.EllipseOverlay;
+import org.apache.commons.math3.analysis.MultivariateFunction;
 import org.apache.commons.math3.optim.InitialGuess;
 import org.apache.commons.math3.optim.MaxEval;
 import org.apache.commons.math3.optim.PointValuePair;
@@ -78,7 +79,6 @@ public class GaussianFit implements StatefulProcessor {
         data.put(outputKey + "x", x);
         data.put(outputKey + "y", y);
         data.put(outputKey + "sigma", sigma);
-
         data.put(outputKey + "overlay_1", new EllipseOverlay(x, y, r + sigma, r + sigma, 0));
         data.put(outputKey + "overlay_2", new EllipseOverlay(x, y, r - sigma, r - sigma, 0));
 
@@ -128,5 +128,46 @@ public class GaussianFit implements StatefulProcessor {
 
     public void setOutputKey(String outputKey) {
         this.outputKey = outputKey;
+    }
+
+    public class GaussianNegLogLikelihood implements MultivariateFunction {
+        private double[] photoncharge;
+        private double[] pixel_x;
+        private double[] pixel_y;
+        private int[] cleaning_pixel;
+
+        /**
+         * @param photoncharge double array containing the photoncharge for each pixel
+         * @param cleaning_pixel int array containing all pixel chids for the pixel that survived cleaning
+         * @param pixel_x double array containing the x coordinates for all pixel
+         * @param pixel_y double array containing the y coordinates for all pixel
+         */
+        public GaussianNegLogLikelihood(double[] photoncharge, int[] cleaning_pixel, double[] pixel_x, double[] pixel_y){
+            this.photoncharge = photoncharge;
+            this.pixel_x = pixel_x;
+            this.pixel_y = pixel_y;
+            this.cleaning_pixel = cleaning_pixel;
+        }
+
+        /**
+         *
+         * @param point a double array with length 4 containing r, x, y, sigma in this order
+         * @return the negative log likelihood at this point for the given data
+         */
+        public double value(double[] point) {
+            double r = point[0];
+            double x = point[1];
+            double y = point[2];
+            double sigma = point[3];
+            double neg_ln_L = 0;
+
+            for (int i = 0; i < cleaning_pixel.length; i++) {
+                int pix = cleaning_pixel[i];
+                double distance = Math.sqrt(Math.pow(pixel_x[pix] - x, 2.0) + Math.pow(pixel_y[pix] - y, 2.0));
+                neg_ln_L += (Math.log(sigma) + Math.pow((distance - r) / sigma, 2)) * photoncharge[pix];
+            }
+
+            return neg_ln_L;
+        }
     }
 }
